requirements.txt
fastapi
uvicorn
pydantic
paddleocr
openai
pytesseract
pdf2image
pillow
python-dotenv
numpy